internal/rest/resources: Reorganize authentication handlers
Moves the authentication handlers out into their own function that is
exported and can be used to facilitate finer control of authentication
per-endpoint.

Additionally adds a new RestrictNotification AccessHandler so that
untrusted endpoints that are re-used for cluster notifications can
restrict those communications only to trusted systems.
