+++ internal/rest/resources/certificates.go
@@ -12,9 +12,9 @@ import (
 	"github.com/canonical/lxd/lxd/response"
 
 	"github.com/canonical/microcluster/client"
-	"github.com/canonical/microcluster/internal/rest/access"
 	"github.com/canonical/microcluster/internal/state"
 	"github.com/canonical/microcluster/rest"
+	"github.com/canonical/microcluster/rest/access"
 	"github.com/canonical/microcluster/rest/types"
 )
