+++ internal/rest/resources/shutdown.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/canonical/lxd/lxd/response"
 
-	"github.com/canonical/microcluster/internal/rest/access"
+	"github.com/canonical/microcluster/rest/access"
 	"github.com/canonical/microcluster/internal/state"
 	"github.com/canonical/microcluster/rest"
 )
